Renamed the module Ably to AblyRealtime
