+++ Spec/AblySpec-Bridging-Header.h
@@ -2,7 +2,7 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-#include "Ably.h"
+#include "AblyRealtime.h"
 #include <asl.h>
 #include "NSObject+TestSuite.h"
 #include "ARTTestUtil.h"
